+++ NEWS.md
@@ -11,7 +11,9 @@ Changes to `read_cell_seg_data`:
 - Use `vroom::vroom()` instead of `readr::read_tsv` for speed and support
   of `col_select`.
   
-- Fix `select_rows` to always return `FALSE` instead of `NA`
+- Add `read_composites` and `read_composite_info` to read and parse
+  metadata in inForm composite images.
+- Fix `select_rows` to always return `FALSE` instead of `NA`.
 
 # phenoptr 0.2.9
 **2020-11-11**